update for cameras
added offset for static secondary picture (on goal) and locked transform to rim because Vector 3 will be different in every scene so i need a reference point to start at

everything seems to work. the 2 orthographic cameras are a little off. not worth the effort it would take to fix